autogen_version/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.6.0
autogen-agentchat==0.6.1
autogen-core==0.6.1
autogen-ext==0.6.1
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
distro==1.9.0
exceptiongroup==1.3.0
fastmcp==2.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.7.0
jiter==0.10.0
jsonref==1.1.0
markdown-it-py==3.0.0
mcp==1.9.3
mdurl==0.1.2
openai==1.86.0
openapi-pydantic==0.5.1
opentelemetry-api==1.34.1
packaging==25.0
pillow==11.2.1
protobuf==5.29.5
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
python-dotenv==1.1.0
python-multipart==0.0.20
qbittorrent-api==2025.5.0
regex==2024.11.6
requests==2.32.4
rich==14.0.0
setuptools==78.1.1
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.7
sse-starlette==2.3.6
starlette==0.47.0
tiktoken==0.9.0
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
wget==3.2
wheel==0.45.1
zipp==3.23.0
